Port TestUpdateNodeWiki
